hr_management_system/requirements.txt
google-generativeai>=0.3.0
langgraph>=0.0.15
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.3.0
jinja2>=3.1.2
tenacity>=8.2.3
uvicorn